Delete Item Cascades deletes associated resource, better validation

+++ src/pages/api/entries/[id].ts
@@ -1,8 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { GetObjectCommand } from "@aws-sdk/client-s3";
+import { GetObjectCommand, DeleteObjectCommand } from "@aws-sdk/client-s3";
 import nextConnect from "next-connect";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { s3Client, spacesConfig } from "../../../server/s3Client";
+import { cloudinary } from "../../../server/cloudinaryClient";
 import { prisma } from "../../../server/db";
 import { getSession } from "next-auth/react";
 
@@ -14,7 +15,6 @@ interface NextRequest extends NextApiRequest {
 
 const handler = nextConnect<NextRequest, NextApiResponse>({
   onError(err, req, res) {
-    console.log(err);
     res.status(500).json({ error: "Route Error", stack: err });
   },
   onNoMatch(_, res) {
@@ -22,7 +22,8 @@ const handler = nextConnect<NextRequest, NextApiResponse>({
   },
 });
 
-// TODO VALIDATE ENTRY BELONG TO USER
+// TODO Separate logic for signed url & detail get
+// GENERATES SIGNED URL AND RETURNS ENTRY DETAILS
 handler.get(async (req, res) => {
   const session = await getSession({ req });
   if (!session) {
@@ -32,7 +33,10 @@ handler.get(async (req, res) => {
     const entryExists = await prisma.entry.findUnique({
       where: { id: id as string },
     });
-    if (entryExists) {
+    const userExists = await prisma.user.findUnique({
+      where: { email: session.user?.email as string },
+    });
+    if (entryExists && userExists && entryExists.userId === userExists.id) {
       const bucketParams = {
         Bucket: $bucket,
         Key: entryExists.file_key as string,
@@ -40,17 +44,20 @@ handler.get(async (req, res) => {
       const signedUrl = await getSignedUrl(
         s3Client,
         new GetObjectCommand(bucketParams),
-        { expiresIn: 15 * 60 }
+        { expiresIn: 5 * 60 }
       );
       res.status(200).json({ url: signedUrl, entryExists });
     } else {
-      res.status(404).json({ error: "Not found" });
+      if (!entryExists) {
+        res.status(404).json({ error: "Not found" });
+      } else if (!userExists || entryExists.userId === userExists.id) {
+        res.status(403).json({ error: "Forbidden" });
+      }
     }
   }
   res.end();
 });
 
-// TODO SENDS DELETE REQ TO ASSOCIATED RESOURCE
 handler.delete(async (req, res) => {
   const session = await getSession({ req });
   if (!session) {
@@ -60,12 +67,27 @@ handler.delete(async (req, res) => {
     const entryExists = await prisma.entry.findUnique({
       where: { id: id as string },
     });
-    if (!entryExists) {
-      res.status(410).json({ error: "Not found" });
-      res.end();
-    } else {
+    const userExists = await prisma.user.findUnique({
+      where: { email: session.user?.email as string },
+    });
+    if (entryExists && userExists && entryExists.userId === userExists.id) {
+      if (entryExists.file_key) {
+        const bucketParams = {
+          Bucket: $bucket,
+          Key: entryExists.file_key as string,
+        };
+        await s3Client.send(new DeleteObjectCommand(bucketParams))
+      }
+      if (entryExists.img_url) {
+        await cloudinary.v2.uploader.destroy(entryExists.img_id as string)
+      }
       await prisma.entry.delete({ where: { id: id as string } });
       res.status(200).json({ message: `Entry ${entryExists!.id} deleted!` });
+    } else {
+      if (!entryExists) {
+        res.status(410).json({ error: "Not found" });
+        res.end();
+      }
     }
   }
   res.end();
